Add files via upload
Testing is deleted and pom file is more than likely unusable. 

Changes to Game class(), mouseInput(), AbstractLevel class has been added, as well as a LevelOne class.
Functionality to come, but this is mainly a refactoring push. The next stage will be to completely break up LevelOne class into an extension of Abstract class. This will be the final push before that possibly game breaking change.